fix clang-tidy issues and add clang-tidy check

+++ src/timerangeeditor.cpp
@@ -32,7 +32,7 @@ void TimeRangeEditor::set_range(const QDateTime& begin, const QDateTime& end)
   m_ui->te_begin->set_time(begin.time());
   m_ui->de_begin->setDate(begin.date());
   m_ui->te_end->set_time(proposed_end.time());
-  m_ui->sp_end_offset->setValue(begin.date().daysTo(proposed_end.date()));
+  m_ui->sp_end_offset->setValue(static_cast<int>(begin.date().daysTo(proposed_end.date())));
   m_ui->cb_has_end->setChecked(end.isValid());
   update();
 }
